remove getAddressFromJson and move ApiAddressException
